- Correction for header miss align - Removed none needed storybook elements - Added footer to header example page - Merged color pages
