+++ public/manager-head.html
@@ -100,6 +100,10 @@
   .header.open ul.dropdown-menu {
     display: block;
   }
+  .Resizer.horizontal,
+  .Resizer.horizontal + div {
+    display: none !important;
+  }
 </style>
 
 <script>
